requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
pypdf
python-dotenv
faiss-cpu
gradio==4.44.1
tiktoken
gradio==4.44.1
gradio-client==1.3.0
huggingface-hub>=0.19.3